fix(ui): allow a newly created task with a sub-board reference to be entered
After adding a task that referenced a sub-board, you were not able to
enter it immediately; you had to go back to tree view and then either
navigate to the newly created sub-board or re-enter parent board to be
able to enter a referenced board. This is not ideal, you want to quickly
navigate to a newly created sub-board.

Previously, after adding a new task that references a sub-board, the
entire tree view was updated by clearing and re-initialing. The issue is
that the column node we are currently viewing has actually been cleared,
so we have to back to the updated tree view and navigate to the updated
column node that contains the new task in order to interact with that
sub-board.