feat(ui): add cycle board task operation
Add operation to move a task to the next column to the right. When
focused column is the last column, wrap around to the first column. When
moving a task to the next column, place the task at the top.

A cycle board task operations avoids the need to move a task by first
deleting and then pasting in the next column. Which can obfuscate the
board task buffer's purpose.