fix(ui): correct function calls to function that calculates a board task's index
Previously, function calls made to function that calculates a board
task's index used right border width for the line width argument. This
led to incorrectly fetching the selected task in a column. For instance,
given a board with multiple columns and say the first column has two
tasks. If the first task was expanded and its description took multiple
lines and then the second task description was expanded, then it was not
possible to collapse the second task since an index out of bounds error
was causing the calculates board task index function to return early.

Those function calls now take pass in line width that it set to be the
width of a single column.