fix(ui): resolve updating tree view when removing multiple columns
Deleting two columns and going back to tree view would only show one of
the columns removed. This required restarting the app to show the
correct tree view. The tree view was being updated by sequentially
searching for the focused column, clearing the found column, and
re-adding the tasks.

The solution was to update the tree view by clearing the board and
re-adding the columns. This is a more computationally expensive
operation since we are updating all columns, including those who have
not been modified.