feat(ui): set task tree view nodes to not be selectable
Set task tree view nodes to not be selectable to improve tree
navigation. It doesn't make sense to set a node to be selectable if it
can't have any children. This is the case since, during tree view
creation, tasks that reference a sub-board use the board for the node
reference instead of the task itself.