refactor(main): change error handling to use log.Printf
Refactor error handling for saving list and board data structures to
files. Instead of exiting the program when error was encountered when
saving the list data structure, simply print the error and continue
execution. This provides the opportunity to attempt to save the board
data structure given that saving the list fails.